+++ vfs.go
@@ -202,6 +202,10 @@ type File interface {
 	// For file:///some/path/to/file.txt, it would return file.txt
 	Name() string
 
+	// Touch creates a zero-length file on the vfs.File if no File exists.  Update File's last modified timestamp.
+	// Returns error if unable to touch File.
+	Touch() error
+
 	// URI returns the fully qualified absolute URI for the File.  IE, s3://bucket/some/path/to/file.txt
 	URI() string
 }